% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irishbabynames_dashboard.R
\name{irishbabynames_dashboard}
\alias{irishbabynames_dashboard}
\title{Launch the Irish Babynames Shiny App}
\usage{
irishbabynames_dashboard()
}
\description{
This function launches the Shiny dashboard included in the irishbabynames package.
}
