% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_similar_names.R
\name{plot_similar_names}
\alias{plot_similar_names}
\title{Internal: Plot phonetically similar names}
\usage{
plot_similar_names(input_name, gender_label, gender_stat, gender_rank_stat)
}
\arguments{
\item{input_name}{Name to find phonetically similar names for}

\item{gender_label}{"Male" or "Female"}

\item{gender_stat}{Statistic string for filtering main dataset}

\item{gender_rank_stat}{Statistic string for filtering ranks}
}
\value{
An interactive ggiraph plot
}
\description{
\code{plot_similar_boysnames()} and \code{plot_similar_girlsnames()} to generate
phonetically similar name plots using the Metaphone algorithm.
}
\keyword{internal}
