% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top10ranks_boys.R
\name{plot_top10_boy_rank_lines}
\alias{plot_top10_boy_rank_lines}
\title{Plot Interactive Line Chart for Top 10 boys' Names by Rank (2001–2024)}
\usage{
plot_top10_boy_rank_lines(data)
}
\arguments{
\item{data}{A dataset like \code{irishbabynames} with columns from 2001 to 2024.}
}
\value{
A \code{girafe} object for interactive display.
}
\description{
This function creates an interactive line plot (using ggiraph) showing how the top 10
boys' names by average rank changed in Ireland between 2001 and 2024.
}
