% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top10_girls.R
\name{plot_top10_girls_occurrence}
\alias{plot_top10_girls_occurrence}
\title{Interactive Plot of Top 10 Girls' Names (1964–2000)}
\usage{
plot_top10_girls_occurrence(data)
}
\arguments{
\item{data}{The full \code{irishbabynames} dataset.}
}
\value{
A \code{girafe} interactive plot object.
}
\description{
This function creates an interactive plot showing the top 10 girls' names
in Ireland from 1964 to 2000 based on number of occurrences.
}
