% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot name trend over years}
\usage{
plot_trend(name_input)
}
\arguments{
\item{name_input}{Name to visualize (character)}
}
\value{
A \code{girafe} interactive ggplot object
}
\description{
Displays the trend of a given name (boys or girls) across years,
showing occurrences and rank as tooltip.
}
