% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_unisex_names.R
\name{plot_unisex_names}
\alias{plot_unisex_names}
\title{Plot Unisex Names Starting with a Letter in a Given Year}
\usage{
plot_unisex_names(start, year)
}
\arguments{
\item{start}{A character letter to filter names starting with it.}

\item{year}{Numeric year (e.g., 2020).}
}
\value{
A plotly bar chart showing unisex names starting with the letter.
}
\description{
Plot Unisex Names Starting with a Letter in a Given Year
}
\examples{
plot_unisex_names("A", 2020)
}
