\name{irishbabynames}
\alias{irishbabynames}
\docType{data}
\title{
IRISH BABY NAMES
}
\description{
 Full baby name data provided by the CSO data from the year 1964 to 2024. This includes all names with at
least 3 occurrences.
}
\usage{data("irishbabynames")}
\format{
  A data frame with 12854 observations on the following 64 variables.
  \describe{
    \item{\code{Statistic}}{a factor with levels \code{Girls Names in Ireland with 3 or More Occurrences Rank} \code{Girls Names in Ireland with 3 or More Occurrences} \code{Boys Names in Ireland with 3 or More Occurrences Rank} \code{Boys Names in Ireland with 3 or More Occurrences}}
    \item{\code{Names}}{Names of for boys and girls.}
    \item{\code{Gender}}{a character vector}
  }
}

\source{
Data obtained from the cso data package.
}

\keyword{datasets}
