% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_similar_boysnames.R
\name{plot_similar_boysnames}
\alias{plot_similar_boysnames}
\title{Plot Phonetically Similar Boys' Names Over Time}
\usage{
plot_similar_boysnames(input_boys_name, max_distance = 1)
}
\arguments{
\item{input_boys_name}{A name string to find phonetically similar boys' names.}

\item{max_distance}{Maximum string distance to consider for name similarity.}
}
\value{
An interactive ggiraph plot.
}
\description{
This function visualizes phonetically similar boys' names based on the Metaphone algorithm.
It shows trends over years using per million births for similar names for boys.
}
