% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_similar_girlsnames.R
\name{plot_similar_girlsnames}
\alias{plot_similar_girlsnames}
\title{Plot Phonetically Similar Girls' Names Over Time}
\usage{
plot_similar_girlsnames(input_name1, max_distance = 1)
}
\arguments{
\item{input_name1}{A name string to find phonetically similar girls' names.}
}
\value{
An interactive ggiraph plot.
}
\description{
This function visualizes phonetically similar girls' names based on the Metaphone algorithm.
It shows trends over years using per million births for similar names for girls.
}
